permissions CRUD + web

+++ app/com/aaabramov/encoded/core/entity/Role.scala
@@ -2,7 +2,7 @@ package com.aaabramov.encoded.core.entity
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Role(name: String, id: Long = 0L)
+case class Role(name: String, id: Int = 0)
 
 object Role {
   implicit val format: OFormat[Role] = Json.format[Role]
